fix: 데이터베이스와 자바의 타입 불일치 해결 및 도커를 사용한 데이터베이스 연결

+++ src/main/java/org/swbe/domain/servicerequest/entity/RequestAssignment.java
@@ -50,5 +50,5 @@ public class RequestAssignment {
   private LocalDateTime endedAt;
 
   @Column(name = "current_marker", insertable = false, updatable = false)
-  private Integer currentMarker;
+  private Byte currentMarker;
 }
